+++ packages/theme/src/client/utils/linkInfo.ts
@@ -1,8 +1,11 @@
-import { ensureEndingSlash, isLinkAbsolute } from "@vuepress/helper/client";
-import { resolveRoute } from "vuepress/client";
+import {
+  ensureEndingSlash,
+  isLinkAbsolute,
+  resolveRoute,
+} from "@vuepress/helper/client";
 
-import type { ArticleInfo, AutoLinkOptions } from "../../shared/index.js";
-import { ArticleInfoType } from "../../shared/index.js";
+import type { AutoLinkOptions, PageInfoData } from "../../shared/index.js";
+import { PageInfo } from "../../shared/index.js";
 
 export const resolvePrefix = (prefix = "", path = ""): string =>
   isLinkAbsolute(path) ? path : `${ensureEndingSlash(prefix)}${path}`;
@@ -14,19 +17,18 @@ export const resolvePrefix = (prefix = "", path = ""): string =>
 export const resolveLinkInfo = (
   item: string,
   preferFull = false,
+  current?: string,
 ): AutoLinkOptions => {
-  const { meta, path, notFound } = resolveRoute<ArticleInfo>(item);
+  const { meta, path, notFound } = resolveRoute<PageInfoData>(item, current);
 
   return notFound
     ? { text: path, link: path }
     : {
         text:
-          !preferFull && meta[ArticleInfoType.shortTitle]
-            ? meta[ArticleInfoType.shortTitle]
-            : meta[ArticleInfoType.title] || path,
+          !preferFull && meta[PageInfo.shortTitle]
+            ? meta[PageInfo.shortTitle]
+            : meta[PageInfo.title] || path,
         link: path,
-        ...(meta[ArticleInfoType.icon]
-          ? { icon: meta[ArticleInfoType.icon] }
-          : {}),
+        ...(meta[PageInfo.icon] ? { icon: meta[PageInfo.icon] } : {}),
       };
 };